+++ src/lib.rs
@@ -25,6 +25,7 @@ pub mod error;
 pub mod eval;
 pub mod fmt_cst;
 pub mod fmt_json;
+pub mod fmt_json_lines;
 pub mod fmt_raw;
 pub mod fmt_rcl;
 pub mod fmt_toml;